python_code/api/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# Google AI and LangChain
google-generativeai>=0.4.1,<0.5.0
langchain-google-genai==0.0.11
langchain==0.1.9

# Vector Database
pinecone-client==3.0.2

# Environment and Utilities
python-dotenv==1.0.1
pandas==2.0.3

# Deployment
runpod==1.3.3

# HTTP Client
httpx==0.27.0

# CORS
starlette==0.36.3

# Logging and Error Handling
traceback-with-variables==2.0.4

# JSON Processing
orjson==3.9.15

# Additional Dependencies
tqdm>=4.65.0
protobuf>=4.22.0
google-api-core>=2.11.0
google-auth>=2.17.0